# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Study Planner API",
    "description": "\n# API Documentation - AI Study Planner\n\nApplication web intelligente pour la génération automatique de plannings d'études personnalisés.\n\n## Fonctionnalités Principales\n\n### 🎓 Gestion Étudiante\n- Authentification JWT\n- Profils étudiants personnalisés\n- Gestion des matières et examens\n- Suivi de progression ECTS\n\n### 📅 Planification Intelligente\n- Génération automatique de plannings\n- Optimisation basée sur IA (Llama 3.2)\n- Contraintes et disponibilités\n- Historique des plannings\n\n### 🏢 Plateforme Super Admin\n- Gestion institutionnelle (universités, programmes, cours)\n- Import/Export en masse (CSV/Excel)\n- Système d'audit complet\n- RBAC (Role-Based Access Control)\n\n### 📊 Analyse & Suivi\n- Analyse de risque académique\n- Priorisation des matières\n- Recommandations personnalisées\n- Tracking ECTS\n\n## Authentification\n\nL'API utilise JWT (JSON Web Tokens) pour l'authentification.\n\n**Obtenir un token:**\n```bash\nPOST /api/v1/auth/login\n{\n  \"email\": \"user@example.com\",\n  \"password\": \"password\"\n}\n```\n\n**Utiliser le token:**\n```\nAuthorization: Bearer <votre_token>\n```\n\n## Environnements\n\n- **Development**: http://localhost:8000\n- **Swagger UI**: http://localhost:8000/docs\n- **ReDoc**: http://localhost:8000/redoc\n\n## Support\n\nPour toute question, consultez la documentation complète dans le README.md\n            ",
...
    "contact": {
...
  "servers": [
...
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register new user",
        "description": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "full_name": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "full_name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "409": {
            "description": "Email already exists"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login user",
        "description": "Authenticate and get JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns JWT token"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}